trying to figured out

+++ app/src/main/java/ca/qc/cstj/consortium/domain/models/Delivery.kt
@@ -6,12 +6,12 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "deliveries")
 data class Delivery (
-    @ColumnInfo(name="eplil") var eplil: Float,
-    @ColumnInfo(name="Awhil") var Awhil: Float,
-    @ColumnInfo(name="vethyx") var vethyx: Float,
-    @ColumnInfo(name="laspyx") var laspyx: Float,
-    @ColumnInfo(name="smiathil") var smiathil: Float,
+    @ColumnInfo(name="eplil") var eplil: Double,
+    @ColumnInfo(name="Awhil") var Awhil: Double,
+    @ColumnInfo(name="vethyx") var vethyx: Double,
+    @ColumnInfo(name="laspyx") var laspyx: Double,
+    @ColumnInfo(name="smiathil") var smiathil: Double,
 
     ) {
-    @PrimaryKey(autoGenerate = true) var idDelivery : Int = 0
+    @PrimaryKey(autoGenerate = true) var idDelivery : Double = 0.0
 }
